requirements.txt
sqlalchemy
faker
google-generativeai
python-dotenv
